backend/requirements.txt
fastapi
uvicorn[standard]
runwayml
google-api-python-client
facebook-sdk
tiktok-api
pytest
